Improve navbar layout. * Refactor styles.
