+++ styles/base/base.scss
@@ -4,7 +4,7 @@
 
 html,
 body {
-  background: variables.$secondary-50;
+  background-color: variables.$secondary-50;
   scroll-behavior: smooth;
   scroll-padding-top: variables.$navbar-height;
   color: variables.$neutral-900;
